refactor(seeding): organize entities by type and create migration
