+++ FastFoodOperator.Api/Data/AppDbContext.cs
@@ -1,5 +1,4 @@
-﻿using FastFoodOperator.Api.Data.Models;
-using FastFoodOperator.Api.Entities;
+﻿using FastFoodOperator.Api.Entities;
 using Microsoft.EntityFrameworkCore;
 
 namespace FastFoodOperator.Api.Data
@@ -158,11 +157,8 @@ namespace FastFoodOperator.Api.Data
             modelBuilder.Entity<ProductVariant>()
                 .Property(pv => pv.PriceModifier)
                 .HasColumnType("decimal(10, 2)");
-
-            if (Environment.GetEnvironmentVariable("ASPNETCORE_ENVIRONMENT") == "Development") 
-            {
-                DatabaseSeeder.Seed(modelBuilder);
-            }
+            
+            DatabaseSeeder.Seed(modelBuilder);
         }
 	}   
 }
