+++ FastFoodOperator.Api/Data/DatabaseSeeder.cs
@@ -1,7 +1,7 @@
 ﻿using FastFoodOperator.Api.Entities;
 using Microsoft.EntityFrameworkCore;
 
-namespace FastFoodOperator.Api.Data.Models
+namespace FastFoodOperator.Api.Data
 {
 	public static class DatabaseSeeder
 	{
@@ -13,28 +13,47 @@ namespace FastFoodOperator.Api.Data.Models
 				new Category { Id = 3, Name = "Sides" }
 			);
 
-			modelBuilder.Entity<Product>().HasData(
-				new Product { Id = 1, Name = "Cheeseburger", Description = "A classic cheeseburger", BasePrice = 5.99m, CategoryId = 1 },
-				new Product { Id = 2, Name = "Coca-Cola", Description = "Refreshing soda", BasePrice = 1.99m, CategoryId = 2 },
-				new Product { Id = 3, Name = "French Fries", Description = "Crispy golden fries", BasePrice = 2.99m, CategoryId = 3 }
-			);
-
-			modelBuilder.Entity<Combo>().HasData(
-				new Combo { Id = 1, Name = "Burger Combo", BasePrice = 8.99m },
-				new Combo { Id = 2, Name = "Big Meal", BasePrice = 12.99m }
-			);
-
 			modelBuilder.Entity<Ingredient>().HasData(
 				new Ingredient { Id = 1, Name = "Lettuce", PriceModifier = 0.50m },
 				new Ingredient { Id = 2, Name = "Tomato", PriceModifier = 0.50m },
 				new Ingredient { Id = 3, Name = "Pickles", PriceModifier = 0.25m }
 			);
 
-			modelBuilder.Entity<ProductVariant>().HasData(
-				new ProductVariant { Id = 1, ProductId = 1, Name = "Double Cheeseburger", Description = "Extra patty", PriceModifier = 2.00m },
-				new ProductVariant { Id = 2, ProductId = 2, Name = "Large Coke", Description = "Bigger size", PriceModifier = 1.00m }
+			modelBuilder.Entity<ProductIngredient>().HasData(
+				new ProductIngredient { ProductId = 1, IngredientId = 1, Required = true },
+				new ProductIngredient { ProductId = 1, IngredientId = 2, Required = true },
+				new ProductIngredient { ProductId = 3, IngredientId = 3, Required = false }
+			);
+			
+			modelBuilder.Entity<Product>().HasData(
+				new Product { Id = 1, Name = "Cheeseburger", Description = "A classic cheeseburger", BasePrice = 5.99m, CategoryId = 1 },
+				new Product { Id = 2, Name = "Coke", Description = "Refreshing soda", BasePrice = 1.99m, CategoryId = 2 },
+				new Product { Id = 3, Name = "French Fries", Description = "Crispy golden fries", BasePrice = 2.99m, CategoryId = 3 }
 			);
 
+			modelBuilder.Entity<ProductVariant>().HasData(
+				new ProductVariant { Id = 1, ProductId = 2, Name = "Small" },
+				new ProductVariant { Id = 2, ProductId = 2, Name = "Medium", PriceModifier = 1.49m },
+				new ProductVariant { Id = 3, ProductId = 2, Name = "Large", PriceModifier = 2.49m },
+				new ProductVariant { Id = 4, ProductId = 3, Name = "Small" },
+				new ProductVariant { Id = 5, ProductId = 3, Name = "Medium", PriceModifier = 1.49m },
+				new ProductVariant { Id = 6, ProductId = 3, Name = "Large", PriceModifier = 2.49m }
+			);
+
+			modelBuilder.Entity<Combo>().HasData(
+				new Combo { Id = 1, Name = "Cheeseburger Combo", BasePrice = 8.99m },
+				new Combo { Id = 2, Name = "Cheeseburger Combo Deluxe", BasePrice = 12.99m }
+			);
+			
+			modelBuilder.Entity<ComboProduct>().HasData(
+				new ComboProduct { ComboId = 1, ProductId = 1 },
+				new ComboProduct { ComboId = 1, ProductId = 2, ProductVariantId = 1 },
+				new ComboProduct { ComboId = 1, ProductId = 3, ProductVariantId = 4 },
+				new ComboProduct { ComboId = 2, ProductId = 1 },
+				new ComboProduct { ComboId = 2, ProductId = 2, ProductVariantId = 3 },
+				new ComboProduct { ComboId = 2, ProductId = 3, ProductVariantId = 6 }
+			);
+			
 			modelBuilder.Entity<Order>().HasData(
 				new Order { Id = 1, OrderNumber = 1001, OrderStatus = true},
 				new Order { Id = 2, OrderNumber = 1002, OrderStatus = false}
@@ -49,19 +68,6 @@ namespace FastFoodOperator.Api.Data.Models
 			modelBuilder.Entity<OrderCombo>().HasData(
 				new OrderCombo { OrderId = 1, ComboId = 1, Quantity = 1 },
 				new OrderCombo { OrderId = 2, ComboId = 2, Quantity = 2 }
-			);
-
-			modelBuilder.Entity<ProductIngredient>().HasData(
-				new ProductIngredient { ProductId = 1, IngredientId = 1, Required = true },
-				new ProductIngredient { ProductId = 1, IngredientId = 2, Required = true },
-				new ProductIngredient { ProductId = 3, IngredientId = 3, Required = false }
-			);
-
-			modelBuilder.Entity<ComboProduct>().HasData(
-				new ComboProduct { ComboId = 1, ProductId = 1 },
-				new ComboProduct { ComboId = 1, ProductId = 2 },
-				new ComboProduct { ComboId = 2, ProductId = 1 },
-				new ComboProduct { ComboId = 2, ProductId = 3 }
 			);
 		}
 	}
